Courses Spit into Components
