+++ app/courses/components/features.ts
@@ -1,0 +1,57 @@
+import { Feature } from './types';
+
+export const commonFeatures: Feature[] = [
+  {
+    title: "LinkedIn Profile",
+    description: "Learn professional LinkedIn strategies and best practices to create an impactful profile that attracts recruiters and opportunities",
+    icon: "Linkedin"
+  },
+  {
+    title: "Resume Review",
+    description: "Comprehensive resume review and optimization to align with industry standards",
+    icon: "FileText"
+  },
+  {
+    title: "CMU Degree Program Insights",
+    description: "In-depth exploration of CMU's degree programs with detailed insights from Program Directors",
+    icon: "GraduationCap"
+  },
+  {
+    title: "CMU Admissions Guide",
+    description: "Direct guidance from CMU admissions committee members on application strategy",
+    icon: "ClipboardCheck"
+  },
+  {
+    title: "Silicon Valley Networking Events",
+    description: "Exclusive access to Silicon Valley tech meetups, conferences, and industry networking events",
+    icon: "Network"
+  },
+  {
+    title: "CMU Student and Faculty Connections",
+    description: "Direct networking opportunities with current CMU students and distinguished faculty members",
+    icon: "Users"
+  }
+];
+
+export const expertFeatures: Feature[] = [
+  {
+    title: "LORs from Microsoft Employees",
+    description: "Receive personalized letters of recommendation for university admissions from Microsoft senior engineers and industry leaders, highlighting your potential for graduate studies",
+    icon: "FileText"
+  },
+  {
+    title: "1:1 Microsoft Executive Mentoring",
+    description: "Exclusive access to three personalized 30-minute mentoring sessions with Microsoft employees from our mentorship network, focusing on career guidance, technical expertise, and professional development",
+    icon: "Users2"
+  },
+  {
+    title: "Microsoft Career Opportunity",
+    description: "Receive a direct internal referral for a position at Microsoft through our professional network, enhancing your application for roles that match your skills and experience",
+    icon: "Target"
+  },
+  {
+    title: "Elite Interview Preparation Program",
+    description: "Comprehensive interview preparation package including three mock interviews with Microsoft engineers, detailed feedback sessions, and personalized strategies for technical and behavioral questions",
+    icon: "Building2"
+  }
+];
